:truck: Moved schema headers introspection to schema.py module
