+++ zds_client/schema.py
@@ -1,5 +1,5 @@
 import logging
-from typing import List, Mapping, Union
+from typing import List, Mapping, Tuple, Union
 from urllib.parse import urlparse
 
 logger = logging.getLogger(__name__)
@@ -85,6 +85,90 @@ def extract_params(url: str, pattern: str) -> dict:
 
 def noop(arg):
     return arg
+
+
+def separate_params(params: List[dict]) -> Tuple[List, List[str]]:
+    """Separate parameters explicitly defined and referenced by `$ref`"""
+    reference_params = []
+    regular_params = []
+    for param in params:
+        if param.get("$ref") is None:
+            regular_params.append(param)
+        else:
+            reference_params.append(param)
+
+    return regular_params, reference_params
+
+
+def filter_header_regular_params(params: list) -> list:
+    return [param for param in params if param["in"] == "header" and param["required"]]
+
+
+def filter_header_reference_params(params: list, spec: dict) -> list:
+    """Filter header parameters which are in definitions referenced with `$ref`"""
+    header_params = []
+
+    for param in params:
+        reference = param.get("$ref")
+        # Local reference case (parameter in specification document)
+        if reference[:2] == "#/":
+            split_path = reference[2:].split("/")
+            tmp_parameter = spec
+            for parent in split_path:
+                tmp_parameter = tmp_parameter.get(parent)
+
+            if tmp_parameter["in"] == "header" and tmp_parameter["required"]:
+                header_params.append(tmp_parameter)
+        # TODO Remote reference case (parameter in a document on the same server)
+        elif "//" not in reference:
+            raise NotImplementedError("To be implemented")
+        # TODO URL reference case (parameter in a document on another server)
+        elif "//" in reference:
+            raise NotImplementedError("To be implemented")
+
+    return header_params
+
+
+def filter_header_params(params: list, spec: dict) -> list:
+    """Extract parameters required for headers"""
+    # Separate the parameters that use references
+    regular_parameters, reference_parameters = separate_params(params)
+
+    # Filter regular and reference parameters
+    header_regular_parameters = filter_header_regular_params(regular_parameters)
+    header_reference_parameters = filter_header_reference_params(
+        reference_parameters, spec
+    )
+    return header_regular_parameters + header_reference_parameters
+
+
+def get_headers(spec: dict, operation: str) -> dict:
+    """
+    Extract required headers and use the default value from the API spec.
+    """
+    headers = {}
+
+    for path, methods in spec["paths"].items():
+        path_parameters = filter_header_params(methods.get("parameters", []), spec)
+        for name, method in methods.items():
+            if name == "parameters":
+                continue
+
+            if method["operationId"] != operation:
+                continue
+
+            method_parameters = filter_header_params(method.get("parameters", []), spec)
+
+            for param in path_parameters + method_parameters:
+                enum = param["schema"].get("enum", [])
+                default = param["schema"].get("default")
+
+                assert (
+                    len(enum) == 1 or default
+                ), "Can't choose an appropriate default header value"
+                headers[param["name"]] = default or enum[0]
+
+    return headers
 
 
 class Schema:
